// src/types/Product.ts
export interface Review {
  id: string;
  username: string;
  rating: number;
  description: string;
}

export interface Product {
  id: string;
  title: string;
  description: string;
  price: number;
  discountedPrice: number;
  image: {
    url: string;
    alt: string;
  };
  rating: number;
  tags: string[];
  reviews: Review[];
}

export interface CartItem {
  product: Product;
  quantity: number;
}

export interface ContactFormData {
  fullName: string;
  subject: string;
  email: string;
  message: string;
}